gradle: multi-project build: separated cardTools into emCardTools module
