+++ applet/build.gradle
@@ -68,6 +68,8 @@ repositories {
 
 // Dependencies for your project
 dependencies {
+    compile project(':emCardTools')
+
     testImplementation 'org.testng:testng:6.1.1'
     testImplementation 'org.slf4j:slf4j-api:1.7.2'
     testImplementation 'org.slf4j:slf4j-log4j12:1.7.2'
