+++ applet/src/test/java/e2e/EmeraldAppletAPDUTest.java
@@ -20,8 +20,8 @@ import static applet.EmeraldProtocol.PASSWORD_SLOT_ID_OFFSET;
 import static applet.EmeraldProtocol.PASSWORD_VALUE_OFFSET;
 import static applet.EmeraldProtocol.aesKeyDevelopmentTODO;
 import applet.SecureChannelManager;
-import cardTools.CardManager;
-import static cardTools.Util.prepareParameterData;
+import emCardTools.CardManager;
+import static emCardTools.Util.prepareParameterData;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 import tests.BaseTest;
