+++ emCardTools/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'java-library'
+}
+
+repositories {
+    jcenter()
+    mavenCentral()
+}
+
+dependencies {
+    compile 'com.klinec:jcardsim:3.0.5.9'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+    api 'com.klinec:jcardsim:3.0.5.9'
+}
