+++ emCardTools/src/main/java/emCardTools/CardManager.java
@@ -29,15 +29,16 @@ SOFTWARE.
 ~~~
 */
 
-package cardTools;
+package emCardTools;
 
 import com.licel.jcardsim.io.CAD;
 import com.licel.jcardsim.io.JavaxSmartCardInterface;
 import javacard.framework.AID;
+import java.util.ArrayList;
 
 import javax.smartcardio.*;
-import java.util.ArrayList;
 import java.util.List;
+
 
 /**
  *
