+++ emCardTools/src/main/java/emCardTools/RunConfig.java
@@ -29,7 +29,7 @@ SOFTWARE.
 ~~~
 */
 
-package cardTools;
+package emCardTools;
 
 /**
  * Applet run configuration.
@@ -42,19 +42,19 @@ public class RunConfig {
     public Class appletToSimulate;
     boolean bReuploadApplet = false;
     byte[] installData = null;
-    
+
     public enum CARD_TYPE {
         PHYSICAL, JCOPSIM, JCARDSIMLOCAL, JCARDSIMREMOTE
     }
 
     public CARD_TYPE testCardType = CARD_TYPE.PHYSICAL;
-    
+
     public static RunConfig getDefaultConfig() {
         RunConfig runCfg = new RunConfig();
         runCfg.targetReaderIndex = 0;
         runCfg.testCardType = CARD_TYPE.PHYSICAL;
         runCfg.appletToSimulate = null;
-        
+
         return runCfg;
     }
