+++ emCardTools/src/main/java/emCardTools/SimulatedCard.java
@@ -29,7 +29,7 @@ SOFTWARE.
 ~~~
 */
 
-package cardTools;
+package emCardTools;
 
 import javax.smartcardio.ATR;
 import javax.smartcardio.Card;
@@ -44,37 +44,37 @@ public class SimulatedCard extends Card {
 
     @Override
     public ATR getATR() {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public String getProtocol() {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public CardChannel getBasicChannel() {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public CardChannel openLogicalChannel() throws CardException {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public void beginExclusive() throws CardException {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public void endExclusive() throws CardException {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
     public byte[] transmitControlCommand(int i, byte[] bytes) throws CardException {
-        throw new UnsupportedOperationException("Not supported yet."); 
+        throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
