+++ settings.gradle
@@ -30,5 +30,6 @@ SOFTWARE.
 */
 
 rootProject.name = 'PV204_smartcards_Emerald'
-include 'applet'
+include ':applet'
+include ':emCardTools'
